requirements.txt
requests==2.31.0
instaloader==4.10.3
openai>=1.40.0
twilio==8.10.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
opencv-python==4.8.1.78
pillow==10.1.0
moviepy==1.0.3
pydantic==2.5.0
pydantic-settings==2.1.0
numpy<2
httpx==0.28.1
firecrawl-py==1.4.0